centralize types
