+++ CHANGELOG.md
@@ -10,9 +10,10 @@
 - Consolidate event handlers, perf improvements
 - Refactor `main.ts` into smaller, single-responsibility modules (`modal`, `stats`, `types`) for improved maintainability.
 - Update test suite to accomodate new project structure.
-- Improve `getWordCount` regex to handle abbreviations properly
-- Create shared services to eliminate code duplication between main plugin and modal
-- Refactor `updateWordCount` method into smaller, single-purpose functions for improved maintainability
+- Improve `getWordCount` regex to handle abbreviations properly.
+- Create shared services to eliminate code duplication between main plugin and modal.
+- Refactor `updateWordCount` method into smaller, single-purpose functions for improved maintainability.
+- Centralize types into `types.ts`.
 
 ### Added
