+++ tests/settings.test.ts
@@ -1,8 +1,8 @@
 import {
 	DEFAULT_SETTINGS,
-	MyPluginSettings,
 	StatBarSettingTab,
 } from "../src/settings";
+import { MyPluginSettings } from "../src/types";
 import { PluginSettingTab } from "obsidian";
 
 describe("Settings Tests", () => {
